fix 9 verified review findings on the raft write path + rebuild cluster sim
High-effort review of the raft wiring surfaced 9 confirmed defects, all
fail-stop crashes, silent data-loss paths, or lost coverage. All fixed:

- deposed-leader local claims: fulfillSubscriptions now gated on
  canWriteLocally() (raft_state==leading && isLeader(), or single-node);
  notifyForFrame skips frames rejected with any error; on step-down
  evictWaitingSubscribers() sends MSG_NOT_LEADER to every waiting worker
  and clears subscriptions. Also: clearState() before rebuildState() on
  re-acquire (the stale in-memory PendingIndex accumulated duplicates
  across step-down cycles and could feed fulfillment).
- fetch payload length is a u32 wire field (was u16: @intCast panicked
  on a default-config 64KiB payload and the old clamp truncated larger
  ones). Encoder + bench/sim/lifecycle decoders updated; SDKs must
  mirror (per-job payload-len in 0x82 is now LE u32).
- proposal cap enforced at the boundary: enqueueRecordedBound() computes
  a provable upper bound on a frame's recorded mutation bytes at decode
  time and rejects with err_too_large / HTTP 413 BEFORE any mutation
  (cluster mode only). The post-commit assert stays as a backstop.
  executeBatch flushes the indexer per-frame in raft mode so each raft
  entry carries exactly its own frame's index mutations (one-op-per-
  entry atomicity for followers).
- max_tick_tokens derived from the worst case (max_frames +
  max_maint_tokens = 448): frame-boundary splitting emits up to one
  token per frame; the old 160 assumed perfect chunk packing.
- clusterHash() restored and enforced inside the HMAC'd raft_net
  handshake, unconditionally (no --cluster-secret still validates —
  the misconfig scenario needs no attacker). Hash now also covers
  persist_completed and purge_retention_ns, which master never hashed;
  mismatched shared params refuse to form a cluster in both directions.
- subscribe overflow at max_waiting_conns rejects with MSG_ERROR before
  any ConnState mutation (was: success response + silent drop = worker
  hangs forever); waiting-conn tracking invariant assert restored.
- indexer nearFull() headroom is the derived per-op maximum (256-job
  batch x 3 records = 768), not the whole buffer: the zero-headroom
  counter-delta term forced a flush after every op, destroying tick
  batching. Companion test no longer tautological.
- fetch send-budget liveness: fulfillSubscriptions only admits a conn
  whose free send buffer fits one max-size job, and .send_done re-arms
  the conn's queues, so a re-pushed job ships when the buffer drains
  (was: stranded until unrelated queue activity). handler_fetch first-
  job path now honest; fetchedJobWireSize layout comment corrected
  (constant 20 is exact: u32 prefix +2 cancels the wrongly-documented
  2-byte id/queue prefixes).
- src/sim/cluster.zig rebuilt for raft (deleted with PBR, never
  ported — the only end-to-end multi-node replication coverage).
  Deterministic single-threaded nodes (KV + Pipeline + raft runtime,
  in-memory transport, shared sim clock), leader-vs-follower KV
  byte-comparison after quiescence. Scenarios: 3-node + 5-node
  replication, multi-queue lifecycle incl. replicated auto-delete,
  live check-quorum failover (exercises eviction + no-divergence),
  follower not-leader rejection, determinism (same seed = same run).
  Single-node sim send buffer now follows main.zig's sizing invariant
  (exact-64KiB buffer made the one-max-job room check silently zero
  out fetch coverage).

Tests: full unit suite green; sim green (single-node fetch/ack traffic
restored, 6 cluster scenarios pass, deterministic across runs).

+++ src/rpc.zig
@@ -284,7 +284,7 @@ pub const BufReader = struct {
         return data;
     }
 
-    pub fn skip(self: *BufReader, n: u16) ParseError!void {
+    pub fn skip(self: *BufReader, n: usize) ParseError!void {
         if (self.pos + n > self.data.len) return error.ShortRead;
         self.pos += n;
     }
